1) Изменено название свойства на BoundingBox 2) В CreateEllipse изменен тип параметра pointCount на Int

+++ flop.net/Model/PolygonBuilder.cs
@@ -59,7 +59,7 @@ namespace flop.net.Model
          return new Polygon(points, true);
       }
   
-      public static Polygon CreateEllipse(Point pointA, Point pointB, double pointCount)
+      public static Polygon CreateEllipse(Point pointA, Point pointB, int pointCount)
       {
          PointCollection points = new PointCollection() { };
          Point center = new Point((pointA.X + pointB.X) / 2, (pointA.Y + pointB.Y) / 2);
